apps/adk-agent/requirements.txt
fastapi
uvicorn
google-genai-agents
pydantic
python-dotenv
requests
